Clean up my confusion
IPMIService was never used. Move IpmiPools into the ipam.go file and
remove unused IPMIService struct.
